+++ CHANGELOG.md
@@ -4,10 +4,11 @@ This repository contains releases of the Power Measurement Toolkit (PMT), a
 comprehensive toolkit for power measurement and monitoring. Below are details
 for each release:
 
-## Unreleased
+## 1.3.1
 
 ### Changed:
 
+- Bugfixes
 - Use instantaneous power measurement in `NVML`
 - Change output of polling mode for `PMT` to show individual measurements